Move progress listener for flicker bug out of asset providers
