+++ plugin/src/software/aws/toolkits/eclipse/amazonq/providers/assets/ChatWebViewAssetProvider.java
@@ -98,18 +98,6 @@ public final class ChatWebViewAssetProvider extends WebViewAssetProvider {
             }
         });
 
-        browser.setVisible(false);
-        browser.addProgressListener(new ProgressAdapter() {
-            @Override
-            public void completed(final ProgressEvent event) {
-                Display.getDefault().asyncExec(() -> {
-                    if (!browser.isDisposed()) {
-                        browser.setVisible(true);
-                    }
-                });
-            }
-        });
-
         browser.setText(getContent().get());
     }
